requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu # remove this line if you have CUDA

torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1

fastapi==0.116.1
langchain==0.3.27
langchain_chroma==0.2.5
langchain_community==0.3.27
langchain_core==0.3.72
langchain_google_genai==2.1.8
langchain_huggingface==0.3.1
langchain_ollama==0.3.6
langgraph==0.5.4
pandas==2.3.1
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.47.0
tqdm==4.67.1
unstructured[all-docs]==0.18.11
sentence-transformers==5.0.0
chromadb==1.0.15
pdfminer.six==20250506